sync_table/requirements.txt
aws-lambda-powertools==2.30.1
instructure-dap-client==0.3.18